pyproject.toml
[tool.poetry]
name = "maplestory-price-track"
version = "0.1.0"
description = ""
authors = ["TonyLin <chiacheng.lin0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
tabulate = "^0.9.0"
requests = "^2.32.3"
cloudscraper = "^1.2.71"
discord-py = "^2.4.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
